trying clean code
