+++ handler/handler.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"be-project/models"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -14,46 +15,13 @@ const (
 	ContentTypeText   = "text/plain; charset=utf-8"
 )
 
-type User struct {
-	UserID      int
-	FullName    string
-	Email       string
-	Password    string
-	PhoneNumber string
-	OfficeID    string
-	RoleID      int
-	Latitude    float32
-	Longitude   float32
-	CurrentDate string
-	AuthToken   string
-}
-
-type Respawns struct {
-	Status  int       `json:"status"`
-	Message string    `json:"message"`
-	Data    Dashboard `json:"data"`
-}
-
-type Dashboard struct {
-	FullName        string  `json:"full_name"`
-	RoleDescription string  `json:"role_description"`
-	OfficeLongitude float32 `json:"office_longitude"`
-	OfficeLatitude  float32 `json:"office_latitude"`
-	CurrentDate     string  `json:"dates"`
-	UserLongitude   float32 `json:"user_longitude"`
-	UserLatitude    float32 `json:"user_latitude"`
-}
-
-type Timesheet struct {
-}
-
 func LandingPage(c *gin.Context) {
 	c.Data(http.StatusOK, ContentTypeHTML, []byte("<h1>ini landing page</h1>"))
 }
 
 func GetDashboard(c *gin.Context) {
 	// collect from db
-	data := &Respawns{} // not yet
+	data := &models.Respon{} // not yet
 
 	c.JSON(http.StatusOK, data)
 }
