+++ models/timesheet.go
@@ -1,0 +1,4 @@
+package models
+
+type Timesheet struct {
+}
